# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/categories:
    get:
      summary: 'Get Categories'
      operationId: getCategories
      description: 'List all the categories.'
      parameters:
        -
          in: query
          name: page
          description: 'Which page to show.'
          example: 12
          required: false
          schema:
            type: integer
            description: 'Which page to show.'
            example: 12
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: aut
                      created_at: '2023-12-01T19:02:42.000000Z'
                    -
                      id: 2
                      name: nesciunt
                      created_at: '2023-12-01T19:02:42.000000Z'
                    -
                      id: 3
                      name: quis
                      created_at: '2023-12-01T19:02:42.000000Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: aut
                        created_at: '2023-12-01T19:02:42.000000Z'
                      -
                        id: 2
                        name: nesciunt
                        created_at: '2023-12-01T19:02:42.000000Z'
                      -
                        id: 3
                        name: quis
                        created_at: '2023-12-01T19:02:42.000000Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: aut
                        created_at:
                          type: string
                          example: '2023-12-01T19:02:42.000000Z'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Not found'
                properties:
                  message:
                    type: string
                    example: 'Not found'
      tags:
        - Categories
      security: []
    post:
      summary: 'POST Categories'
      operationId: pOSTCategories
      description: 'Create new category'
      parameters: []
      responses: {  }
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Name of the category.'
                  example: '"Clothing"'
              required:
                - name
      security: []
...